semi_app/semiproj/requirements.txt
Django
slackclient
django-crispy-forms
djangorestframework
mysqlclient
django-concurrency
requests
djangogrpcframework
grpcio
grpcio-tools
googleapis-common-protos
celery
redis
django-celery-results
django-celery-beat
django-redis